Move validator
