+++ src/API/WesternStatesWater.WestDaat.Contracts.Client/Requests/Admin/OrganizationMemberAddRequestValidator.cs
@@ -20,19 +20,3 @@ public class OrganizationMemberAddRequestValidator : AbstractValidator<Organizat
             .WithMessage("Role not allowed.");
     }
 }
-
-public class UserProfileUpdateRequestValidator : AbstractValidator<UserProfileUpdateRequest>
-{
-    public UserProfileUpdateRequestValidator()
-    {
-        RuleFor(x => x.FirstName).NotEmpty().MaximumLength(255);
-
-        RuleFor(x => x.LastName).NotEmpty().MaximumLength(255);
-
-        RuleFor(x => x.State).NotEmpty().MaximumLength(2);
-
-        RuleFor(x => x.Country).NotEmpty().MaximumLength(50);
-
-        RuleFor(x => x.PhoneNumber).NotEmpty().MaximumLength(50);
-    }
-}
